refactor(v2-ui): delete reference-review page — review derives from projection (spec §3.5)
