+++ extralit-frontend/translation/en.js
@@ -141,21 +141,9 @@ export default {
     loadError: "Could not load the extraction table.",
   },
   review: {
-    title: "Review",
-    submit: "Submit",
-    saveDraft: "Save draft",
-    discard: "Discard",
-    suggestion: "Suggestion",
-    response: "Response",
     agent: "Agent",
     score: "Score",
-    context: "Context",
     notApplicable: "Not applicable in this schema version",
-    orphanedValues: "Values from removed questions (read-only, not resubmitted)",
-    loadError: "Could not load review for this reference.",
-    submitted: "Response submitted",
-    draftSaved: "Draft saved",
-    discarded: "Response discarded",
   },
   settings: {
     title: "Dataset settings",
