+++ extralit-frontend/v2/di/di.ts
@@ -7,13 +7,11 @@ import { SchemaRepository } from "~/v2/infrastructure/repositories/SchemaReposit
 import { V2RecordRepository } from "~/v2/infrastructure/repositories/V2RecordRepository";
 import { AnnotationRepository } from "~/v2/infrastructure/repositories/AnnotationRepository";
 import { ProjectionRepository } from "~/v2/infrastructure/repositories/ProjectionRepository";
-import { GetReferenceReviewUseCase } from "~/v2/domain/usecases/get-reference-review-use-case";
 import { GetWorkspaceProjectionUseCase } from "~/v2/domain/usecases/get-workspace-projection-use-case";
 import { useExtractions } from "~/v2/infrastructure/storage/ExtractionsStorage";
 import { SubmitReferenceReviewUseCase } from "~/v2/domain/usecases/submit-reference-review-use-case";
 import { SaveReviewDraftUseCase } from "~/v2/domain/usecases/save-review-draft-use-case";
 import { DiscardReviewUseCase } from "~/v2/domain/usecases/discard-review-use-case";
-import { useReferenceReviews } from "~/v2/infrastructure/storage/ReferenceReviewsStorage";
 import { useSchemas } from "~/v2/infrastructure/storage/SchemasStorage";
 import { GetSchemasUseCase } from "~/v2/domain/usecases/get-schemas-use-case";
 import { GetSchemaSettingsUseCase } from "~/v2/domain/usecases/get-schema-settings-use-case";
@@ -46,15 +44,6 @@ export const loadV2DependencyContainer = (nuxtApp: NuxtAppLike) => {
     register(ProjectionRepository).withDependency(useAxios).build(),
     register(GetWorkspaceProjectionUseCase).withDependencies(ProjectionRepository, useExtractions).build(),
 
-    register(GetReferenceReviewUseCase)
-      .withDependencies(
-        ProjectionRepository,
-        SchemaRepository,
-        V2RecordRepository,
-        AnnotationRepository,
-        useReferenceReviews
-      )
-      .build(),
     register(SubmitReferenceReviewUseCase).withDependency(AnnotationRepository).build(),
     register(SaveReviewDraftUseCase).withDependency(AnnotationRepository).build(),
     register(DiscardReviewUseCase).withDependency(AnnotationRepository).build(),
